+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The **patch** part changes is incremented if multiple releases happen the same m
 * evomaintenance: allow missing API endpoint if APi is disabled
 * java: use default JRE package when version is not specified
 * listupgrade: better detection for PostgreSQL
+* listupgrade: sort/uniq of packages/services lists in email template
 * lxc-solr: detect the real partition options
 * lxc-solr: download URL according to Solr Version
 * lxc-solr: set homedir and port at install